Sum of the range between two numbers
We did two cases, the firts case was about when a and b are equal, and we returned a. And the second case we identify te maximum and minimum then we got the range between these numbers, and then we added this numbers, in the end we returned the sum.